requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.5
python-multipart==0.0.6
python-dotenv==1.0.0
typing_extensions==4.7.1
